fix(migration): check db connection type before drop foreign key of group_id to prevent exception in sqlite database which not supports drop foreign

+++ database/migrations/2025_10_14_073346_add_cascade_on_delete_attribute_to_group_id_foreign_key_of_todos_table.php
@@ -12,9 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('todos', function (Blueprint $table) {
-            $table->dropForeign('tasks_group_id_foreign');
-            $table->dropColumn('group_id');
-            $table->foreignId('group_id')->constrained()->cascadeOnDelete();
+            if (Schema::getConnection()->getDriverName() !== 'sqlite') {
+                $table->dropForeign('tasks_group_id_foreign');
+                $table->dropColumn('group_id');
+                $table->foreignId('group_id')->constrained()->cascadeOnDelete();
+            } 
         });
     }
